backend/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database - Using asyncpg instead of psycopg2 for Python 3.13 compatibility
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.14.0

# Redis
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP requests
httpx==0.28.0
requests==2.32.3

# Google Gemini AI
google-generativeai==0.8.3

# GitHub API
PyGithub==2.5.0

# Utilities
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Development
pytest==8.3.3
pytest-asyncio==0.24.0